requirements.txt
fastapi
uvicorn
jinja2
pymongo
pydantic
random
openai
google.generativeai